Review remarks. Moved enums to separate files. Cosmetic changes. I2C address constants renamed
