+++ src/devices/Amg88xx/Register.cs
@@ -2,8 +2,6 @@
 // The .NET Foundation licenses this file to you under the MIT license.
 // See the LICENSE file in the project root for more information.
 
-using System;
-
 namespace Iot.Device.Amg88xx
 {
     /// <summary>
@@ -805,149 +803,5 @@ namespace Iot.Device.Amg88xx
         /// Pixel 64 value register (high byte)
         /// </summary>
         T64H = 0xff
-    }
-
-    /// <summary>
-    /// Defines the status bits of the status register (addr: 0x04)
-    /// </summary>
-    [Flags]
-    public enum StatusFlagBit : byte
-    {
-        /// <summary>
-        /// Interrupt occured
-        /// </summary>
-        INTF = 1,
-
-        /// <summary>
-        /// Temperature output overflow occured for one or more pixel
-        /// </summary>
-        OVF_IRS = 2,
-
-        /// <summary>
-        /// Thermistor output overflow occured
-        /// Note: the bit is only menthioned in early versions of the reference specification.
-        /// It is not clear whether this is a specification error or a change in a newer
-        /// revision of the sensor.
-        /// </summary>
-        OVF_THS = 3
-    }
-
-    /// <summary>
-    /// Defines the status bits of the status clear register (addr: 0x05)
-    /// </summary>
-    [Flags]
-    public enum StatusClearBit : byte
-    {
-        /// <summary>
-        /// Interrupt occured
-        /// </summary>
-        INTCLR = 1,
-
-        /// <summary>
-        /// Temperature output overflow occured for one or more pixel
-        /// </summary>
-        OVFCLR = 2,
-
-        /// <summary>
-        /// Thermistor output overflow occured
-        /// </summary>
-        OVFTHCLR = 3
-    }
-
-    /// <summary>
-    /// Defines the bit(s) of the moving average mode register (addr: 0x07)
-    /// </summary>
-    public enum MovingAverageModeBit : byte
-    {
-        /// <summary>
-        /// Twice moving average mode bit
-        /// </summary>
-        MAMOD = 5
-    }
-
-    /// <summary>
-    /// Defines the bit of the frame rate register (addr: 0x02)
-    /// </summary>
-    public enum FrameRateBit : byte
-    {
-        /// <summary>
-        /// Frame rate mode bit (not set: 10fps, set: 1fps)
-        /// </summary>
-        FPS = 1
-    }
-
-    /// <summary>
-    /// Defines the operating modes of the power control register (addr: 0x00)
-    /// </summary>
-    public enum OperatingMode : byte
-    {
-        /// <summary>
-        /// normal mode
-        /// </summary>
-        Normal = 0x00,
-
-        /// <summary>
-        /// sleep mode
-        /// </summary>
-        Sleep = 0x10,
-
-        /// <summary>
-        /// stand-by mode, 10s intermittence
-        /// </summary>
-        StandBy10 = 0x21,
-
-        /// <summary>
-        /// stand-by mode, 60s intermittence
-        /// </summary>
-        StandBy60 = 0x20,
-    }
-
-    /// <summary>
-    /// Defines the reset types of the reset register (addr: 0x01)
-    /// </summary>
-    public enum ResetType : byte
-    {
-        /// <summary>
-        /// flag reset (clear all flags and interrupt flag registers)
-        /// </summary>
-        Flag = 0x30,
-
-        /// <summary>
-        /// initial reset (set all registers to defaults)
-        /// </summary>
-        Initial = 0x3f
-    }
-
-    /// <summary>
-    /// Defines the interrupt modes of the interrupt control register (addr: 0x03)
-    /// </summary>
-    public enum InterruptMode : byte
-    {
-        /// <summary>
-        /// The specification does not give any details on this mode
-        /// </summary>
-        DifferenceMode,
-
-        /// <summary>
-        /// An interrupt occures if any pixel exceed the upper or lower limit as given in the
-        /// interrupt level register
-        /// </summary>
-        AbsoluteMode
-    }
-
-    /// <summary>
-    /// Defines the interrupt controls bits of the interrupt control register (addr: 0x03)
-    /// </summary>
-    public enum InterruptModeBit : byte
-    {
-        /// <summary>
-        /// Interrupt output bit (INT pin, 0: inactive, 1: active)
-        /// </summary>
-        INTEN = 0,
-
-        /// <summary>
-        /// Interrupt mode bit (0: difference mode, 1: absolute mode)
-        /// </summary>
-        INTMODE = 1
     }
 }
